+++ renderer/units/Unicode Inspector/import.js
@@ -89,6 +89,10 @@ module.exports.start = function (context)
                     let maxLength = charactersInput.maxLength;
                     if (text.length > maxLength)
                     {
+                        if (/[\uD800-\uDBFF]/.test (text[maxLength - 1]))   // Unpaired high surrogate
+                        {
+                            maxLength = maxLength - 1;
+                        }
                         text = text.substring (0, maxLength);
                     }
                     charactersInput.value = text;